Fix a crash when opening dir with auto-save on
Fix this bug: when auto-save is enabled, and the previous default save
dir is invalid now (say it has been removed), the program will crash
when user try to open an image dir.

This bug is due to wrong order of variable initialization and poor edge
case protection. In the above mentioned situation, self.file_path is set
to None and never initialized or checked again, until it is passed to
os.path.splitext and fails.